Refactor client code (#54)
